Re-org postgres wallet storage code
