+++ experimental/plugins/postgres_storage/postgres-cli-test.sh
@@ -1,0 +1,2 @@
+cd ../../../cli && \
+RUST_BACKTRACE=1 LD_LIBRARY_PATH=../experimental/plugins/postgres_storage/target/debug/ cargo run ../experimental/plugins/postgres_storage/cli_ps_test.txt
