+++ experimental/plugins/postgres_storage/src/libindy/wallet.rs
@@ -1,5 +1,5 @@
 
-use api::ErrorCode;
+use libindy::ErrorCode;
 
 use libc::c_char;
 use utils::callbacks;
