+++ experimental/plugins/postgres_storage/src/wql/mod.rs
@@ -1,3 +1,5 @@
 
 pub mod language;
+pub mod query;
 pub mod storage;
+pub mod transaction;
